Move Healthpack stuff to utility_items.py
