+++ src/test/test_dungeon_controller.py
@@ -5,7 +5,7 @@ from unittest.mock import Mock
 from pygame.math import Vector2
 
 import mods
-from mods import HealthPackObject
+from items.utility_items import HealthPackObject
 from items.bullet_weapons import PistolObject, ShotgunObject
 from src.test.pygame_mock import initialize_pygame
 from src.test.testing_utilities import make_dungeon_controller, make_player
